W231: surplus 64-bit arithmetic and the over-window population — 13 windows cleared, nd -3881
Largest nd reduction since wave 223, and the most over-windows cleared in any wave. 0 MATCHes lost,
0 new over-window, **3,072 bytes of code retail does not contain removed from the tree.**

The wave targeted two measurable classes and, crucially, their intersection. Every `(long)` cast and
`u64` local the original did not have makes b210 emit a `dsll32`/`dsra32` pair — 8 bytes of arithmetic
the program never performed, since PS2 `long` is 64-bit and the decompiler guesses it freely. And an
object larger than its retail window contains invented logic by definition, so the excess is a
measurement. **`btlAction 00290be0` was the flagship prediction: 140 bytes over window and 30 surplus
64-bit sites, which at 8 bytes each was more than enough to explain it. It cleared** — 16828 -> 16672
against a 16688 window, nd11848 -> 11197.

**Thirteen windows cleared:** `mt_evtMenu 003760a0` (the tree's largest excess, **+1380 -> inside**,
15476 -> 14032, nd11573 -> 10238), `btlAction 00290be0`, `y_runtime 00458cb0` (+420 -> inside, -484
bytes), `00457a40`, `00438010`, `mdlManager 00313490` and `0031b820`, `mdlFile 003542a0`,
`h_camp 0011bbb0`, `game_support_late 0018c150`, `_h_camp_main 0013be50`, `fclShopMisc 00405f70`,
`k_field 001ba3d0`.

**Three general causes of object bloat were isolated, and they are more useful than the type work:**

1. **A re-derived value that the original retained.** `mt_evtMenu 003760a0` recomputed `row * 12` at
   every use; the original kept the incoming parameter plus ONE persistent derived local. The
   decompiler cannot distinguish a re-derived value from a retained one, so it re-derives everywhere.
   This single mechanism removed 1444 bytes, and A231Camp then used it to clear `_h_camp_main
   0013be50`, reporting it as decisive.
2. **A hand-computed record address where a real record exists.** Typing `workData` as the recovered
   `FldRootWork` stopped MWCC caching a synthetic `work+0x10` pointer (`k_field 001ba3d0` nd282 -> 228,
   window cleared exactly). A231Yaj applied wave 228's address-subtraction test to discover
   `DAT_0086e6a0` is a *second base for the same* stride-0x1c0 `YRuntimeUnitRow`, worth 484 bytes at
   `00458cb0`. The wave-228 headers keep compounding.
3. **Genuinely invented code.** An extra draw call retail does not make plus a stale branch where
   retail assigns the state directly (`fclShopMisc 00405f70`); an invented bounded-switch default and
   duplicated `i == 5` branches (`h_camp 0011bbb0`); a redundant inner NULL guard already guaranteed by
   its enclosing condition (`mdlManager 00313490`).

**On the type work itself, the important discovery is that the defect runs in BOTH directions.** A
too-NARROW signed type is just as costly: at `h_maestro 00114e70` the packed colour channels were
declared `s8`, producing sign-extension pairs *and* signed-u8-to-float correction paths. Declaring
`r`/`g`/`b`/`a` as `u32` with byte masks removed all of it (nd806 -> 677). Meanwhile the wrong wide
type usually enters from a **callee whose result is declared 64-bit** and then propagates into the
comparisons and the loop increment — fix it at the call and the downstream casts become deletable
(`mt_evtMain 003638e0`, `mt_evtMisc 0038b960`, `mdlFile 0031b820`'s seven `(long)(void*)mdlSearch(...)`
conversions, `itfMesManager 003a53b0` nd55 -> 22, and one saved child pointer being carried through
`unsigned __int128`).

**The discipline in the declines is the other half of this result.** Agents declined roughly seventy
targets, and the dominant reason was the right one: **retail contains the same wide operations**, so
our declaration was already correct. Also declined as genuine: the Camp draw code's packed `u64`
coordinate pairs (twice, at the two heaviest 64-bit sites in the tree), packed float-coordinate pairs
required by a callee ABI, a `u64` packet UID, and `_pextlb` operands. A231Btl reverted a probe that
*shrank* an object because the retail diff proved the fallback branch and both state-setting calls it
would have deleted are really there — declining to fit a window by deleting real code is exactly
right. Two more were reverted purely for ending up over-window despite better metrics (`fclShopMisc
0040fb90`, `mdlManager 0031c1d0`, the latter after a traced 112-byte reduction left it 28 bytes short).

Method note: my census regex over-counted, flagging `long` tokens in functions whose fndiff shows no
surplus at all. A231Panel caught this and declined eight targets on that basis. **The fndiff is
authoritative; a textual scan is only a lead.** Census: `build/wip/W231_types.md`.

+++ src/Event/mt_evtMain.c
@@ -673,7 +673,7 @@ void FUN_00362f20(int param_1);
 void FUN_00362fa0(int param_1);
 void FUN_00362ff0(int param_1);
 void FUN_00363030(int *param_1,u16 *param_2,u32 param_3,u8 *param_4);
-void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,u64 param_5);
+void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,int param_5);
 void FUN_003648e0(int param_1,u32 param_2,int param_3);
 void FUN_00364a30(u32 param_1,int param_2);
 void FUN_00364e40(u32 param_1,int param_2,int param_3);
@@ -3537,7 +3537,7 @@ void FUN_00363030(int *param_1,u16 *param_2,u32 param_3,u8 *param_4)
 // FUN_003638E0 NONMATCHING
 
 
-void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,u64 param_5)
+void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,int param_5)
 
 
 
@@ -3549,7 +3549,7 @@ void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,u64 param_5)
 
   u32 lVar3;
 
-  long lVar4;
+  int lVar4;
 
   u32 uVar5;
 
@@ -3571,7 +3571,7 @@ void FUN_003638e0(int *param_1,int param_2,int param_3,u32 param_4,u64 param_5)
 
   u16 *puVar14;
 
-  long lVar15;
+  int lVar15;
 
   long lVar16;
